docs/極秘_特許申請書類/極秘　優先順位1位　次世代エンタープライズAI統合プラットフォーム v3.0 統合エンタープライズAI・セキュリティプラットフォーム v3.0-GISM/app/services/distributed_processing.py
"""
分散処理サービス
- Apache Spark: 大規模データ分散処理
- Apache Kafka: ストリーミングデータ処理
- Ray: 分散機械学習
"""
import os
import asyncio
from typing import List, Dict, Any, Optional
import logging

try:
    from pyspark.sql import SparkSession
    SPARK_AVAILABLE = True
except ImportError:
    SPARK_AVAILABLE = False
    logging.warning("PySpark not available. Spark functionality will be limited.")

try:
    from kafka import KafkaProducer, KafkaConsumer
    KAFKA_AVAILABLE = True
except ImportError:
    KAFKA_AVAILABLE = False
    logging.warning("kafka-python not available. Kafka functionality will be limited.")

try:
    import ray
    RAY_AVAILABLE = True
except ImportError:
    RAY_AVAILABLE = False
    logging.warning("Ray not available. Ray functionality will be limited.")

logger = logging.getLogger(__name__)


class DistributedProcessingService:
    """分散処理サービス"""
    
    def __init__(self):
        self.spark = None
        self.kafka_producer = None
        self.kafka_consumer = None
        self.ray_initialized = False
        
        # Spark初期化
        if SPARK_AVAILABLE:
            self._init_spark()
        
        # Kafka初期化
        if KAFKA_AVAILABLE:
            self._init_kafka()
        
        # Ray初期化
        if RAY_AVAILABLE:
            self._init_ray()
    
    def _init_spark(self):
        """Spark初期化"""
        try:
            self.spark = SparkSession.builder \
                .appName("MultimodalAI-Platform") \
                .config("spark.sql.warehouse.dir", "/tmp/spark-warehouse") \
                .getOrCreate()
            logger.info("Spark initialized successfully")
        except Exception as e:
            logger.error(f"Failed to initialize Spark: {e}")
            self.spark = None
    
    def _init_kafka(self):
        """Kafka初期化"""
        try:
            kafka_bootstrap_servers = os.getenv(
                "KAFKA_BOOTSTRAP_SERVERS",
                "localhost:9092"
            )
            self.kafka_producer = KafkaProducer(
                bootstrap_servers=kafka_bootstrap_servers
            )
            logger.info("Kafka producer initialized successfully")
        except Exception as e:
            logger.error(f"Failed to initialize Kafka: {e}")
            self.kafka_producer = None
    
    def _init_ray(self):
        """Ray初期化"""
        try:
            if not ray.is_initialized():
                ray.init(ignore_reinit_error=True)
                self.ray_initialized = True
                logger.info("Ray initialized successfully")
        except Exception as e:
            logger.error(f"Failed to initialize Ray: {e}")
            self.ray_initialized = False
    
    def is_available(self) -> bool:
        """サービス利用可能性チェック"""
        return (
            (SPARK_AVAILABLE and self.spark is not None) or
            (KAFKA_AVAILABLE and self.kafka_producer is not None) or
            (RAY_AVAILABLE and self.ray_initialized)
        )
    
    async def process_large_scale_data(
        self,
        data_source: str
    ) -> Dict[str, Any]:
        """
        Sparkを使用した大規模データ処理
        """
        if not SPARK_AVAILABLE or self.spark is None:
            return {
                "status": "error",
                "message": "Spark is not available"
            }
        
        try:
            # データ読み込み
            df = self.spark.read.parquet(data_source) if data_source.endswith('.parquet') else \
                 self.spark.read.csv(data_source, header=True, inferSchema=True)
            
            # データ処理例
            processed_df = df.select("*").limit(1000)  # サンプル処理
            
            # 結果取得
            count = processed_df.count()
            
            return {
                "status": "success",
                "record_count": count,
                "message": f"Processed {count} records using Spark"
            }
        except Exception as e:
            logger.error(f"Spark processing error: {e}")
            return {
                "status": "error",
                "message": str(e)
            }
    
    async def stream_data(
        self,
        topic: str,
        callback: Optional[callable] = None
    ) -> Dict[str, Any]:
        """
        Kafkaを使用したストリーミングデータ処理
        """
        if not KAFKA_AVAILABLE or self.kafka_producer is None:
            return {
                "status": "error",
                "message": "Kafka is not available"
            }
        
        try:
            # メッセージ送信例
            message = {
                "topic": topic,
                "status": "streaming",
                "timestamp": asyncio.get_event_loop().time()
            }
            
            self.kafka_producer.send(
                topic,
                value=str(message).encode('utf-8')
            )
            
            return {
                "status": "success",
                "message": f"Message sent to topic: {topic}",
                "data": message
            }
        except Exception as e:
            logger.error(f"Kafka streaming error: {e}")
            return {
                "status": "error",
                "message": str(e)
            }
    
    async def distributed_training(
        self,
        model_config: Dict[str, Any]
    ) -> Dict[str, Any]:
        """
        Rayを使用した分散学習
        """
        if not RAY_AVAILABLE or not self.ray_initialized:
            return {
                "status": "error",
                "message": "Ray is not available"
            }
        
        try:
            # 分散学習タスク定義例
            @ray.remote
            def train_model_worker(data_chunk):
                # ここに実際の学習ロジックを実装
                return {"accuracy": 0.95, "loss": 0.05}
            
            # 分散実行
            results = ray.get([
                train_model_worker.remote(chunk)
                for chunk in range(4)  # 4つのワーカー
            ])
            
            return {
                "status": "success",
                "message": "Distributed training completed",
                "results": results
            }
        except Exception as e:
            logger.error(f"Ray training error: {e}")
            return {
                "status": "error",
                "message": str(e)
            }
    
    async def process_time_series(
        self,
        time_series_data: List[float]
    ) -> Dict[str, Any]:
        """
        時系列データ処理（分散処理を使用）
        """
        if not SPARK_AVAILABLE or self.spark is None:
            # Sparkが利用できない場合は通常処理
            return {
                "status": "success",
                "mean": sum(time_series_data) / len(time_series_data),
                "count": len(time_series_data)
            }
        
        try:
            # Spark DataFrameに変換
            from pyspark.sql import Row
            rows = [Row(value=val) for val in time_series_data]
            df = self.spark.createDataFrame(rows)
            
            # 集計処理
            result = df.agg({
                "value": "mean",
                "value": "count"
            }).collect()[0]
            
            return {
                "status": "success",
                "mean": result["avg(value)"],
                "count": result["count(value)"]
            }
        except Exception as e:
            logger.error(f"Time series processing error: {e}")
            return {
                "status": "error",
                "message": str(e)
            }

